[chores] Add Todo Utils.js, which was ignored
